Revert "Move get_location_path to utils"
5711a43afb3094b09290bfd9677aa49464f290b9
